feat: added authentication method to all routes

+++ src/routes/memories.ts
@@ -3,8 +3,17 @@ import { z } from "zod";
 import { prisma } from "../lib/prisma";
 
 export async function memoriesRoutes(server: FastifyInstance) {
-  server.get("/", async () => {
+  server.addHook("preHandler", async (request) => {
+    // Aqui o jwtVerify vai validar se o usuário está autenticado para acessar a rota
+    await request.jwtVerify();
+  });
+
+  server.get("/", async (request) => {
     const memories = await prisma.memory.findMany({
+      where: {
+        // Lista apenas as memórias do usuário logado
+        id: request.user.sub,
+      },
       orderBy: {
         createdAt: "asc",
       },
@@ -18,7 +27,7 @@ export async function memoriesRoutes(server: FastifyInstance) {
     });
   });
 
-  server.get("/:id", async (request) => {
+  server.get("/:id", async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().cuid(),
     });
@@ -30,10 +39,17 @@ export async function memoriesRoutes(server: FastifyInstance) {
         id,
       },
     });
+
+    if (!memory?.isPublic && memory?.userId !== request.user.sub) {
+      return reply.status(401).send({
+        message: "Memória não está publica",
+      });
+    }
+
     return memory;
   });
 
-  server.post("/memories", async (request) => {
+  server.post("/compose", async (request) => {
     const bodySchema = z.object({
       content: z.string(),
       coverUrl: z.string(),
@@ -47,14 +63,14 @@ export async function memoriesRoutes(server: FastifyInstance) {
         content,
         coverUrl,
         isPublic,
-        userId: "clhrzs9440000827c3a8n7vdq",
+        userId: request.user.sub,
       },
     });
 
     return memory;
   });
 
-  server.put("/memories/:id", async (request) => {
+  server.put("/:id", async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().cuid(),
     });
@@ -65,6 +81,18 @@ export async function memoriesRoutes(server: FastifyInstance) {
     });
     const { id } = paramsSchema.parse(request.params);
     const { content, coverUrl, isPublic } = bodySchema.parse(request.body);
+
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    });
+
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send({
+        message: "Memoria não pertence ao usuário logado",
+      });
+    }
 
     const updatedMemory = await prisma.memory.update({
       where: {
@@ -80,18 +108,28 @@ export async function memoriesRoutes(server: FastifyInstance) {
     return updatedMemory;
   });
 
-  server.delete("/memories/:id", async (request) => {
+  server.delete("/memories/:id", async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().cuid(),
     });
     const { id } = paramsSchema.parse(request.params);
 
-    const deletedMemory = await prisma.memory.delete({
+    let memory = await prisma.memory.findUniqueOrThrow({
       where: {
         id,
       },
     });
 
-    return deletedMemory;
+    if (memory.userId !== request.user.sub) {
+      return reply.status(401).send({
+        message: "Memoria não pertence ao usuário logado",
+      });
+    }
+
+    await prisma.memory.delete({
+      where: {
+        id,
+      },
+    });
   });
 }
